Move win32 key counter into separate module
